feat(connect,builder-tools): move connect vite config logic to builder tools
