--- apps/connect/client.config.ts
@@ -1,71 +1,0 @@
-export default {
-  meta: [
-    {
-      tag: "meta",
-      attrs: {
-        property: "og:title",
-        content: "Connect",
-      },
-    },
-    {
-      tag: "meta",
-      attrs: {
-        property: "og:type",
-        content: "website",
-      },
-    },
-    {
-      tag: "meta",
-      attrs: {
-        property: "og:url",
-        content: "https://apps.powerhouse.io/powerhouse/connect/",
-      },
-    },
-    {
-      tag: "meta",
-      attrs: {
-        property: "og:description",
-        content:
-          "Navigate your organisation’s toughest operational challenges and steer your contributors to success with Connect. A navigation, collaboration and reporting tool for decentralised and open organisation.",
-      },
-    },
-    {
-      tag: "meta",
-      attrs: {
-        property: "og:image",
-        content:
-          "https://cf-ipfs.com/ipfs/bafkreigrmclndf2jpbolaq22535q2sw5t44uad3az3dpvkzrnt4lpjt63e",
-      },
-    },
-    {
-      tag: "meta",
-      attrs: {
-        name: "twitter:card",
-        content: "summary_large_image",
-      },
-    },
-    {
-      tag: "meta",
-      attrs: {
-        name: "twitter:image",
-        content:
-          "https://cf-ipfs.com/ipfs/bafkreigrmclndf2jpbolaq22535q2sw5t44uad3az3dpvkzrnt4lpjt63e",
-      },
-    },
-    {
-      tag: "meta",
-      attrs: {
-        name: "twitter:title",
-        content: "Connect",
-      },
-    },
-    {
-      tag: "meta",
-      attrs: {
-        name: "twitter:description",
-        content:
-          "Navigate your organisation’s toughest operational challenges and steer your contributors to success with Connect. A navigation, collaboration and reporting tool for decentralised and open organisation.",
-      },
-    },
-  ],
-};
