+++ apps/connect/tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../tsconfig.options.json",
   "compilerOptions": {
-    "noEmit": true,
+    "outDir": "./lib",
     "lib": ["ESNext", "DOM", "DOM.Iterable"],
     "types": [
       "node",
@@ -9,13 +9,9 @@
       "vite-plugin-svgr/client",
       "@types/wicg-file-system-access"
     ],
-    "paths": {
-      "@powerhousedao/connect": ["./src/index.ts"],
-      "@powerhousedao/connect/config": ["./src/connect.config.ts"]
-    }
   },
-  "include": ["**/*", "src/i18n/locales/en.json"],
-  "exclude": ["dist", "node_modules", "storybook-static"],
+  "include": ["**/*", "src/i18n/locales/en.json", "./package.copy.json"],
+  "exclude": ["dist", "lib", "node_modules", "storybook-static"],
   "references": [
     {
       "path": "../../packages/document-model"
