+++ packages/builder-tools/connect-utils/index.ts
@@ -1,8 +1,10 @@
 export * from "./constants.js";
 export * from "./helpers.js";
+export * from "./vite-config.js";
 export * from "./vite-plugins/base.js";
 export * from "./vite-plugins/document-models-hmr.js";
 export * from "./vite-plugins/editors-hmr.js";
 export * from "./vite-plugins/external-packages.js";
 export * from "./vite-plugins/importmap.js";
+export * from "./vite-plugins/ph-external-packages.js";
 export * from "./vite-plugins/studio.js";
