+++ packages/builder-tools/package.json
@@ -33,10 +33,16 @@
   },
   "devDependencies": {
     "@storybook/types": "^8.5.2",
+    "@tailwindcss/vite": "^4.1.13",
     "@types/react": "^18.3.18",
     "@types/validator": "^13.12.2",
+    "@vitejs/plugin-react": "^5.0.3",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "vite": "^7.1.7",
+    "vite-plugin-html": "^3.2.2",
+    "vite-plugin-node-polyfills": "^0.24.0",
+    "vite-plugin-svgr": "^4.5.0"
   },
   "dependencies": {
     "@codemirror/autocomplete": "^6.18.6",
@@ -97,7 +103,6 @@
     "uuid": "^11.1.0",
     "validator": "^13.15.0",
     "viem": "^2.28.0",
-    "vite": "^7.1.3",
     "vite-envs": "^4.6.0",
     "vite-plugin-node-polyfills": "^0.24.0",
     "zod": "^3.24.3"
